Guess aggressively at the end of the game (relatively ad-hoc)
This just tweaks the "risky plays" part of `decide_wrapped()` to take
much more risk once the deck ran out.

Also, once the deck runs out, if there is exactly one card that might be
playable, we can deduce that the card is playable and play it. I just
hardcoded this (but we can take out that chuck of code and only incur a
tiny penalty).